+++ experimental/packages/opentelemetry-exporter-metrics-otlp-http/test/metricsHelper.ts
@@ -30,7 +30,8 @@ import {
   AggregationTemporality,
   ExplicitBucketHistogramAggregation,
   MeterProvider,
-  MetricReader
+  MetricReader,
+  View
 } from '@opentelemetry/sdk-metrics-base';
 import {
   IExportMetricsServiceRequest,
@@ -61,6 +62,11 @@ class TestMetricReader extends MetricReader {
   }
 }
 
+export const HISTOGRAM_AGGREGATION_VIEW = new View({
+  aggregation: new ExplicitBucketHistogramAggregation([0, 100]),
+  instrumentName: 'int-histogram',
+});
+
 const defaultResource = Resource.default().merge(new Resource({
   service: 'ui',
   version: 1,
@@ -78,8 +84,8 @@ export async function collect() {
   return (await reader.collect())!;
 }
 
-export function setUp() {
-  meterProvider = new MeterProvider({ resource: defaultResource });
+export function setUp(views?: View[]) {
+  meterProvider = new MeterProvider({ resource: defaultResource, views });
   reader = new TestMetricReader();
   meterProvider.addMetricReader(
     reader
@@ -124,7 +130,6 @@ export function mockDoubleCounter(): Counter {
 }
 
 
-
 export function mockObservableCounter(
   callback: (observableResult: ObservableResult) => void,
   name = 'double-observable-counter'
@@ -158,18 +163,7 @@ export function mockObservableUpDownCounter(
 }
 
 export function mockHistogram(): Histogram {
-  const name = 'int-histogram';
-
-  meterProvider.addView({
-    aggregation: new ExplicitBucketHistogramAggregation([0, 100])
-  },
-  {
-    instrument: {
-      name: name
-    }
-  });
-
-  return meter.createHistogram(name, {
+  return meter.createHistogram('int-histogram', {
     description: 'sample histogram description',
     valueType: ValueType.INT,
   });